refactor: extract mockRpcNodeRequest
