+++ apps/cowswap-e2e-tests/src/mocks/ethGetCode.ts
@@ -1,4 +1,6 @@
-import type { BrowserContext, Route } from '@playwright/test'
+import { mockRpcNodeRequest } from '../support/mockRpcNodeRequest'
+
+import type { BrowserContext } from '@playwright/test'
 
 export interface EthGetCodeMock {
   /** Override the bytecode reported for `address` — e.g. a non-`'0x'` value to simulate a
@@ -29,35 +31,7 @@ interface JsonRpcEntry {
 export function installEthGetCode(context: BrowserContext): EthGetCodeMock {
   const overrides = new Map<string, string>()
 
-  void context.route('**/*', async (route: Route) => {
-    const request = route.request()
-    if (request.method() !== 'POST') return route.fallback()
-
-    let body: JsonRpcEntry | JsonRpcEntry[]
-    try {
-      body = request.postDataJSON() as JsonRpcEntry | JsonRpcEntry[]
-    } catch {
-      return route.fallback()
-    }
-
-    const entries = Array.isArray(body) ? body : [body]
-    if (!entries.some((entry) => entry?.method === 'eth_getCode')) return route.fallback()
-
-    if (entries.every((entry) => entry?.method === 'eth_getCode')) {
-      const payload = entries.map((entry) => ({
-        jsonrpc: '2.0',
-        id: entry.id,
-        result: resolveCode(entry, overrides),
-      }))
-      return route.fulfill({
-        status: 200,
-        contentType: 'application/json',
-        body: JSON.stringify(Array.isArray(body) ? payload : payload[0]),
-      })
-    }
-
-    return fulfillFromUpstream(route, entries, overrides)
-  })
+  mockRpcNodeRequest(context, 'eth_getCode', (entry) => resolveCode(entry, overrides))
 
   return {
     set(address, code) {
@@ -66,37 +40,6 @@ export function installEthGetCode(context: BrowserContext): EthGetCodeMock {
     reset() {
       overrides.clear()
     },
-  }
-}
-
-/**
- * A mixed batch alongside something else this mock doesn't own — patch only the `eth_getCode`
- * slots and merge with the real response for the rest, with the same defensive try/catch as the
- * allowances/SocketVerifier mocks so a flaky real upstream can't take the whole batch down.
- */
-async function fulfillFromUpstream(
-  route: Route,
-  entries: JsonRpcEntry[],
-  overrides: Map<string, string>,
-): Promise<void> {
-  try {
-    const upstream = await route.fetch()
-    const upstreamBody = (await upstream.json()) as JsonRpcEntry | JsonRpcEntry[]
-    const upstreamEntries = Array.isArray(upstreamBody) ? upstreamBody : [upstreamBody]
-    const byId = new Map(entries.map((entry) => [entry.id, entry]))
-
-    const payload = upstreamEntries.map((entry) => {
-      const original = byId.get((entry as JsonRpcEntry).id)
-      if (!original || original.method !== 'eth_getCode') return entry
-      return { jsonrpc: '2.0', id: original.id, result: resolveCode(original, overrides) }
-    })
-    await route.fulfill({
-      status: 200,
-      contentType: 'application/json',
-      body: JSON.stringify(Array.isArray(upstreamBody) ? payload : payload[0]),
-    })
-  } catch {
-    await route.fallback()
   }
 }
